introduce shared nano states, rename nano values to nano states
