+++ app/canvas/canvas.tsx
@@ -30,17 +30,20 @@ import {
 import { useUpdateStyle } from "./shared/style";
 import { useActiveElementTracking } from "./shared/active-element";
 import { WrapperComponentDev } from "./features/wrapper-component";
-import {
-  rootInstanceContainer,
-  useRootInstance,
-  useBreakpoints,
-} from "./shared/nano-values";
 import { useSync } from "./shared/sync";
 import { useManageProps } from "./shared/props";
 import {
   useHandleBreakpoints,
   useInitializeBreakpoints,
 } from "./shared/breakpoints";
+import {
+  rootInstanceContainer,
+  useBreakpoints,
+  useRootInstance,
+} from "~/shared/nano-states";
+import { registerContainers } from "./shared/immerhin";
+
+registerContainers();
 
 const useElementsTree = () => {
   const [rootInstance] = useRootInstance();
