+++ app/canvas/features/wrapper-component/editable/use-content-editable.tsx
@@ -5,7 +5,7 @@ import ObjectId from "bson-objectid";
 import { type ChildrenUpdates, type Instance } from "@webstudio-is/sdk";
 import { type OnChangeChildren } from "~/shared/tree-utils";
 import { primitives } from "~/shared/component";
-import { useSelectedInstance } from "~/canvas/shared/nano-values";
+import { useSelectedInstance } from "~/canvas/shared/nano-states";
 import { useEditable, type Content } from "./use-editable";
 import { useToolbar } from "./use-toolbar";
