+++ app/canvas/shared/style.ts
@@ -2,7 +2,8 @@ import store from "immerhin";
 import { useSubscribe } from "@webstudio-is/sdk";
 import { type StyleUpdates } from "~/shared/component";
 import { setInstanceStyleMutable } from "~/shared/tree-utils";
-import { rootInstanceContainer, useSelectedInstance } from "./nano-values";
+import { useSelectedInstance } from "./nano-states";
+import { rootInstanceContainer } from "~/shared/nano-states";
 
 export const useUpdateStyle = () => {
   const [selectedInstance] = useSelectedInstance();
