+++ app/designer/designer.tsx
@@ -10,12 +10,10 @@ import { SidebarLeft } from "./features/sidebar-left";
 import { Inspector } from "./features/inspector";
 import { CanvasIframe } from "./shared/canvas-iframe";
 import {
-  useIsPreviewMode,
-  useRootInstance,
   useSelectedInstanceData,
   useSyncStatus,
   useCanvasWidth,
-} from "./shared/nano-values";
+} from "./shared/nano-states";
 import { Topbar } from "./features/topbar";
 import designerStyles from "./designer.css";
 import { Breadcrumbs } from "./features/breadcrumbs";
@@ -27,6 +25,7 @@ import {
 import { useReadCanvasRect, Workspace } from "./features/workspace";
 import { usePublishShortcuts } from "./shared/shortcuts";
 import { type SyncStatus } from "~/shared/sync";
+import { useIsPreviewMode, useRootInstance } from "~/shared/nano-states";
 
 export const links = () => {
   return [
