+++ app/designer/features/breakpoints/breakpoints-editor.tsx
@@ -4,7 +4,7 @@ import { type Breakpoint, type Publish } from "@webstudio-is/sdk";
 import { Button, TextField, Flex, Text } from "~/shared/design-system";
 import { PlusIcon, TrashIcon } from "~/shared/icons";
 import ObjectId from "bson-objectid";
-import { useBreakpoints } from "~/designer/shared/nano-values";
+import { useBreakpoints } from "~/shared/nano-states";
 
 type BreakpointEditorItemProps = {
   breakpoint: Breakpoint;
