+++ app/designer/features/breakpoints/width-setting.tsx
@@ -1,12 +1,11 @@
 import { useMemo } from "react";
 import {
   useCanvasWidth,
-  useIsPreviewMode,
   useSelectedBreakpoint,
-  useBreakpoints,
-} from "~/designer/shared/nano-values";
+} from "~/designer/shared/nano-states";
 import { Slider, Text, Flex } from "~/shared/design-system";
 import { sort } from "~/shared/breakpoints";
+import { useBreakpoints, useIsPreviewMode } from "~/shared/nano-states";
 
 // Doesn't make sense to allow resizing the canvas lower/higher than this.
 export const minWidth = 50;
