+++ app/designer/features/sidebar-left/panels/components/custom-drag-layer.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { createPortal } from "react-dom";
 import { useDragLayer, type XYCoord } from "react-dnd";
 import { type Instance } from "@webstudio-is/sdk";
-import { useCanvasRect, useZoom } from "~/designer/shared/nano-values";
+import { useCanvasRect, useZoom } from "~/designer/shared/nano-states";
 import { ComponentThumb } from "./component-thumb";
 
 const layerStyles = {
