+++ app/designer/features/topbar/share.tsx
@@ -9,7 +9,7 @@ import {
 } from "~/shared/design-system";
 import { Share1Icon } from "~/shared/icons";
 import { type Project } from "@webstudio-is/sdk";
-import { useIsShareDialogOpen } from "../../shared/nano-values";
+import { useIsShareDialogOpen } from "../../shared/nano-states";
 
 type ShareButtonProps = { path: string; project: Project };
