+++ app/designer/features/topbar/sync-status.tsx
@@ -1,6 +1,6 @@
 import { Flex, keyframes, AccessibleIcon } from "~/shared/design-system";
 import { CheckIcon, DotsHorizontalIcon } from "~/shared/icons";
-import { useSyncStatus } from "../../shared/nano-values";
+import { useSyncStatus } from "../../shared/nano-states";
 
 const iconSize = 15;
