+++ app/designer/shared/nano-states/index.ts
@@ -1,19 +1,13 @@
 import { createValueContainer, useValue } from "react-nano-state";
-import { type Breakpoint, type Instance } from "@webstudio-is/sdk";
+import { type Breakpoint } from "@webstudio-is/sdk";
 import { SelectedInstanceData } from "~/shared/component";
 import { type SyncStatus } from "~/shared/sync";
-
-const rootInstanceContainer = createValueContainer<Instance | undefined>();
-export const useRootInstance = () => useValue(rootInstanceContainer);
 
 const selectedInstanceDataContainer = createValueContainer<
   SelectedInstanceData | undefined
 >();
 export const useSelectedInstanceData = () =>
   useValue(selectedInstanceDataContainer);
-
-const isPreviewModeContainer = createValueContainer<boolean>(false);
-export const useIsPreviewMode = () => useValue(isPreviewModeContainer);
 
 const isShareDialogOpenContainer = createValueContainer<boolean>(false);
 export const useIsShareDialogOpen = () => useValue(isShareDialogOpenContainer);
@@ -27,9 +21,6 @@ const selectedBreakpointContainer = createValueContainer<
 >();
 export const useSelectedBreakpoint = () =>
   useValue(selectedBreakpointContainer);
-
-export const breakpointsContainer = createValueContainer<Array<Breakpoint>>([]);
-export const useBreakpoints = () => useValue(breakpointsContainer);
 
 const zoomContainer = createValueContainer<number>(100);
 export const useZoom = () => useValue(zoomContainer);
